--- a/src/sql/upgrade-69.sql
+++ b/src/sql/upgrade-69.sql
@@ -1,0 +1,1 @@
+create index IndexBuildOutputsPath on BuildOutputs using hash(path);